% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessedData.R
\name{PreprocessedData}
\alias{PreprocessedData}
\title{Construct a \link{PreprocessedData} object}
\usage{
PreprocessedData(
  sample,
  qc,
  thresholds = NULL,
  sample_outliers,
  metabolite_pass,
  sample_filter_log,
  metabolite_filter_log,
  params = list()
)
}
\arguments{
\item{sample}{A \link{ConcentrationData}.}

\item{qc}{A \link{QCData}.}

\item{thresholds}{A \link{ThresholdData} or \code{NULL}.}

\item{sample_outliers}{Logical vector, length \verb{nrow(@sample@assay)}.}

\item{metabolite_pass}{Logical vector, length \verb{ncol(@sample@assay)}.}

\item{sample_filter_log}{Data.frame, one row per sample.}

\item{metabolite_filter_log}{Data.frame, one row per metabolite.}

\item{params}{List of parameters used by \code{preprocess()}.}
}
\value{
A \link{PreprocessedData} object.
}
\description{
Construct a \link{PreprocessedData} object
}
