% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-diagnostics.R
\name{qc_permanova}
\alias{qc_permanova}
\title{QC PERMANOVA (kit effect)}
\usage{
qc_permanova(prep)
}
\arguments{
\item{prep}{A \link{PreprocessedData} object.}
}
\value{
The \code{adonis2} result table (a data.frame subclass).
}
\description{
Quantify how much of the QC-sample variation is explained by \code{kit_id}
via \code{vegan::adonis2} on Euclidean distance over log-transformed values.
}
